% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPoints.R
\name{matchPoints}
\alias{matchPoints}
\title{Calculates the total goals of the match}
\usage{
matchPoints(df)
}
\arguments{
\item{df}{Match data.}
}
\value{
A data frame containing the final scores, and points for the ladder.
}
\description{
\code{matchPoints} calculates final match goals and score difference.
}
\details{
\code{matchPoints()} treats \code{goal_from_zone1} as one point and
\code{goal_from_zone2} as two points, matching the current super shot era.
}
