% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superNetballR.R
\docType{package}
\name{superNetballR-package}
\alias{superNetballR}
\alias{superNetballR-package}
\title{Functions getting and manipulating Super Netball data.}
\description{
This package provides functions to easily download and manipulate data from super netball matches.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://craigmoyle.github.io/superNetballR_updated}
  \item \url{https://github.com/craigmoyle/superNetballR_updated}
  \item Report bugs at \url{https://github.com/craigmoyle/superNetballR_updated/issues}
}

}
\author{
\strong{Maintainer}: Steve Lane \email{lane.s@unimelb.edu.au}

}
\keyword{internal}
